src/main/java/guru/springframework/mapper/NotesCommandToNotes.java
package guru.springframework.mapper;

import guru.springframework.dto.NotesDTO;
import guru.springframework.model.Notes;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

/**
 * Created by jt on 6/21/17.
 */
@Component
public class NotesCommandToNotes implements Converter<NotesDTO, Notes> {

    @Synchronized
    @Nullable
    @Override
    public Notes convert(NotesDTO source) {
        if(source == null) {
            return null;
        }

        final Notes notes = new Notes();
        notes.setId(source.getId());
        notes.setRecipeNotes(source.getRecipeNotes());
        return notes;
    }
}
